Use specific symbols/structs instead of macros (#105)
Given that hdf5r minimum hdf5 version is 1.8.13, it is safe to use the
1.8+-only versions of the various functions/structs that were previously
addressed via macros. This includes H5Ewalk, H5E_auto_t,
H5E{get,set}_auto and H5Tget_array_dims.

This is usually not a problem when compiling against your everyday hdf5
library, but it becomes an issue when the target hdf5 library was
compiled with --with-default-api-version=v16. This causes the macros to
map by default to the *1 version of the functions/structs, breaking the
code.

An alternative solution was to add the necessary macro definitions in
the Makevars.in file to force the usage of the newer API, but
compatibility with 1.6 is not really required to begin with.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>